+++ client/src/css/form.css
@@ -1,29 +1,15 @@
- .login {
-	 overflow: hidden;
-	 background-color: white;
+.login {
+	display: flex;
+	flex-direction: column;
+	width: 25rem;
+	background-color: gray;
 	 padding: 40px 30px 30px 30px;
 	 border-radius: 10px;
-	 position: absolute;
-	 top: 20%;
-	 left: 10%;
-	 width: 400px;
-	 box-shadow: 5px 10px 10px rgba(2, 128, 144, 0.2);
+	 box-shadow: 5px 10px 10px 10px rgba(2, 128, 144, 0.2);
 }
- 
- .login::before {
-	 left: 40%;
-	 bottom: -130%;
-	 background-color: rgba(69, 105, 144, 0.15);
-}
- .login::after {
-	 left: 35%;
-	 bottom: -125%;
-	 background-color: rgba(2, 128, 144, 0.2);
-}
- .login > input {
-	 font-family: "Asap", sans-serif;
-	 display: block;
-	 border-radius: 5px;
+
+.form-inputs {
+	border-radius: 5px;
 	 font-size: 16px;
 	 background: white;
 	 width: 100%;
@@ -31,29 +17,17 @@
 	 padding: 10px 10px;
 	 margin: 15px -10px;
 }
- .login > button {
-	 font-family: "Asap", sans-serif;
-	 cursor: pointer;
-	 color: #fff;
-	 font-size: 16px;
-	 text-transform: uppercase;
-	 width: 100px;
-	 border: 0;
-	 padding: 10px 0;
-	 margin-top: 10px;
-	 margin-left: -5px;
-	 border-radius: 5px;
-	 background-color: rgba(244, 91, 105, 1);
+
+.form-submit {
+	text-transform: uppercase;
+	width: 100px;
+	border: 0;
+	padding: 10px 0;
+	margin-top: 10px;
+	margin-left: -5px;
+	border-radius: 5px;
 }
- .login > button:hover {
-	 background-color: rgba(242, 67, 83, 1);
+
+.forms-row {
+	height: 43rem;
 }
- a {
-	 text-decoration: none;
-	 color: rgba(255, 255, 255, 0.6);
-	 position: absolute;
-	 right: 10px;
-	 bottom: 10px;
-	 font-size: 12px;
-}
-
